fix(evm): removed redundant (gas_price as u128) cast

+++ apex-sdk-evm/src/provider.rs
@@ -114,7 +114,7 @@ impl CoreProvider for EvmProvider {
             }
         };
 
-        let fee = (gas_price as u128) * gas_limit;
+        let fee = gas_price * gas_limit;
         Ok(fee)
     }
